add pronunciation assessment

+++ db/migrate/20190417155142_create_user_lessons.rb
@@ -1,9 +1,11 @@
+# frozen_string_literal: true
+
 class CreateUserLessons < ActiveRecord::Migration[5.2]
   def change
     create_table :user_lessons do |t|
       t.references :user, foreign_key: true
       t.references :lesson, foreign_key: true
-      t.float :point
+      t.integer :point
 
       t.timestamps
     end
